requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-widget-tweaks==1.5.0
gunicorn==23.0.0
lxml==6.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-decouple==3.8
python-docx==1.2.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
whitenoise==6.9.0
